refactor: centralize user dependency
